# JU-4: Studio performance fixes